refactored the test into an independent module and made the agent executable for manual testing
